use TikZ source instead of PDF for logo in LaTeX
The logo PDF was exported by Figma, but it was not rendered correctly
by some readers. For example, the bubble icon was not rendered at all
in macOS Safari's renderer (although iOS Safari works fine...)?

There was also warnings in the LuaLaTeX log saying the PDF version of
that logo is too high to be fully compatible.

And it's obviously a bad thing to keep track of too many binary files in
a git repository.

So we will use the TikZ source instead. `\globalscale` has been
calibrated to match earlier-generated documents (so they don't have to
be rebuilt again).

Signed-off-by: Vu Van Dung <joulev.vvd@yahoo.com>