remove all tests. okay hear me out: read the commit description below
I agree that tests are necessary. Very necessary even. Without tests god
knows how much we would have to struggle during Orbital. Let alone testing
is a requirement for Orbital higher levels as well.

But this is no Orbital, and I don't think the repo will see much changes
after this, going to the future. Considering how the whole repo has been
virtually rewritten from scratch in v2, the testing mechanism and most test
cases have to be rewritten as well. I don't think it's worth the effort to
do so as the repo won't be that active.

If you somehow stumble upon this issue looking for our Orbital code, see
https://github.com/joulev/ezkomment/tree/v1.0.0, which is truly the
Orbital code unmodified – not the one you're seeing.

Signed-off-by: Vu Van Dung <me@joulev.dev>